+++ VoxelGame/Logic/Blocks/Misc/BedBlock.cs
@@ -81,7 +81,7 @@ namespace VoxelGame.Logic.Blocks
             }
         }
 
-        protected override BoundingBox GetBoundingBox(int x, int y, int z, byte data)
+        protected override BoundingBox GetBoundingBox(int x, int y, int z, uint data)
         {
             bool isBase = (data & 0b1) == 1;
             Orientation orientation = (Orientation)((data & 0b0_0110) >> 1);
@@ -122,10 +122,10 @@ namespace VoxelGame.Logic.Blocks
             return new BoundingBox(new Vector3(0.5f, 0.3125f, 0.5f) + new Vector3(x, y, z), new Vector3(0.5f, 0.125f, 0.5f), legs);
         }
 
-        public override uint GetMesh(BlockSide side, byte data, out float[] vertices, out int[] textureIndices, out uint[] indices, out TintColor tint, out bool isAnimated)
+        public override uint GetMesh(BlockSide side, uint data, out float[] vertices, out int[] textureIndices, out uint[] indices, out TintColor tint, out bool isAnimated)
         {
             bool isHead = (data & 0b1) == 1;
-            int orientation = (data & 0b0_0110) >> 1;
+            int orientation = (int)((data & 0b0_0110) >> 1);
             BlockColor color = (BlockColor)((data & 0b1_1000) >> 3);
 
             if (isHead)
@@ -222,7 +222,7 @@ namespace VoxelGame.Logic.Blocks
             }
         }
 
-        protected override bool Destroy(PhysicsEntity? entity, int x, int y, int z, byte data)
+        protected override bool Destroy(PhysicsEntity? entity, int x, int y, int z, uint data)
         {
             bool isHead = (data & 0b1) == 1;
 
@@ -261,7 +261,7 @@ namespace VoxelGame.Logic.Blocks
             }
         }
 
-        protected override void EntityInteract(PhysicsEntity entity, int x, int y, int z, byte data)
+        protected override void EntityInteract(PhysicsEntity entity, int x, int y, int z, uint data)
         {
             bool isHead = (data & 0b1) == 1;
 
@@ -271,33 +271,33 @@ namespace VoxelGame.Logic.Blocks
 
                     isHead = !isHead;
 
-                    Game.World.SetBlock(this, (byte)(data + 0b0_1000 & 0b1_1111), x, y, z);
-                    Game.World.SetBlock(this, (byte)((data + 0b0_1000 & 0b1_1111) ^ 0b0_0001), x, y, z - (isHead ? 1 : -1));
+                    Game.World.SetBlock(this, data + 0b0_1000 & 0b1_1111, x, y, z);
+                    Game.World.SetBlock(this, (data + 0b0_1000 & 0b1_1111) ^ 0b0_0001, x, y, z - (isHead ? 1 : -1));
                     break;
 
                 case Orientation.East:
 
-                    Game.World.SetBlock(this, (byte)(data + 0b0_1000 & 0b1_1111), x, y, z);
-                    Game.World.SetBlock(this, (byte)((data + 0b0_1000 & 0b1_1111) ^ 0b0_0001), x - (isHead ? 1 : -1), y, z);
+                    Game.World.SetBlock(this, data + 0b0_1000 & 0b1_1111, x, y, z);
+                    Game.World.SetBlock(this, (data + 0b0_1000 & 0b1_1111) ^ 0b0_0001, x - (isHead ? 1 : -1), y, z);
                     break;
 
                 case Orientation.South:
 
-                    Game.World.SetBlock(this, (byte)(data + 0b0_1000 & 0b1_1111), x, y, z);
-                    Game.World.SetBlock(this, (byte)((data + 0b0_1000 & 0b1_1111) ^ 0b0_0001), x, y, z - (isHead ? 1 : -1));
+                    Game.World.SetBlock(this, data + 0b0_1000 & 0b1_1111, x, y, z);
+                    Game.World.SetBlock(this, (data + 0b0_1000 & 0b1_1111) ^ 0b0_0001, x, y, z - (isHead ? 1 : -1));
                     break;
 
                 case Orientation.West:
 
                     isHead = !isHead;
 
-                    Game.World.SetBlock(this, (byte)(data + 0b0_1000 & 0b1_1111), x, y, z);
-                    Game.World.SetBlock(this, (byte)((data + 0b0_1000 & 0b1_1111) ^ 0b0_0001), x - (isHead ? 1 : -1), y, z);
+                    Game.World.SetBlock(this, data + 0b0_1000 & 0b1_1111, x, y, z);
+                    Game.World.SetBlock(this, (data + 0b0_1000 & 0b1_1111) ^ 0b0_0001, x - (isHead ? 1 : -1), y, z);
                     break;
             }
         }
 
-        internal override void BlockUpdate(int x, int y, int z, byte data, BlockSide side)
+        internal override void BlockUpdate(int x, int y, int z, uint data, BlockSide side)
         {
             if (side == BlockSide.Bottom && Game.World.GetBlock(x, y - 1, z, out _)?.IsSolidAndFull != true)
             {
